cleanup jest config
