--- jest.config.js
@@ -1,14 +1,0 @@
-const config = {
-  testEnvironment: 'node',
-  clearMocks: true,
-  restoreMocks: true,
-  resetMocks: true,
-
-  moduleDirectories: [
-    "node_modules",
-    "app/frontend"
-  ]
-};
-
-export default config;
-
