+++ variants/frontend-react/app/frontend/tests/setupExpectEachTestHasAssertions.js
@@ -1,0 +1,5 @@
+/**
+ * Sets up before each test an expectation for there to be assertions.
+ */
+// eslint-disable-next-line jest/require-top-level-describe
+beforeEach(() => expect.hasAssertions());
